requirements.txt
requests
psycopg2-binary
transformers
torch
tqdm
pandas
numpy
python-dotenv
pydantic
pyahocorasick
concurrent-log-handler
streamlit